Move hook to list view folder, move useQuery out of custom hook and return query from hook
